requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
openai>=1.0.0
groq>=0.4.1
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.0.0
